requirements.txt
pyyaml
cdp-py